+++ Transit/Transit/Services/QuickActionService.swift
@@ -1,0 +1,8 @@
+#if os(iOS)
+import Foundation
+
+@Observable
+final class QuickActionService {
+    var pendingNewTask = false
+}
+#endif